Rename project, update to new API and upgrade deps
The new name is 'dvlopt.mbus' as it is more representative.

The JMbus library as been upgraded to version 3.0.1.

The new API consists of similar functions but focused on data and
clojure semantics :

  - Information is represented as maps with namespaced keywords.
  - Namespaced keywords are all specified using clojure.spec.
  - Some namespaced keywords have defined default values.
  - Exceptions are catched and wrapped in maps presenting some context.

Specs related to Meter-Bus in knew are in the 'dvlopt.mbus' namespace
whereas specs related to wired Meter-Bus are specifically located in the
'dvlopt.mbus.wired' namespace. Similar scheme shall be used for wireless
Meter-Bus in the future.